Apply style template to go back to tabs

+++ Genv-experimental/src/main/java/org/genvisis/one/JL/QuantileMathCommonsNormalization.java
@@ -11,17 +11,18 @@ import org.genvisis.common.Array;
 import com.google.common.primitives.Doubles;
 
 public class QuantileMathCommonsNormalization {
-	private static final RankingAlgorithm COV_RANKER_TIE = new NaturalRanking(NaNStrategy.FAILED, TiesStrategy.AVERAGE);
+	private static final RankingAlgorithm COV_RANKER_TIE = new NaturalRanking(NaNStrategy.FAILED,
+																																						TiesStrategy.AVERAGE);
 
 	/**
 	 * Quantile normalize a double[][] double[probes][sample]
 	 *
-	 * @param rawData
-	 *            matrix containing expression/methylation data
+	 * @param rawData matrix containing expression/methylation data
 	 */
 	public static void quantilenormalize(double[][] rawData) {
 		System.out.println("\nPerforming quantile normalization:");
-		// Calculate the average expression, when per sample all raw expression levels have been ordered:
+		// Calculate the average expression, when per sample all raw expression levels have been
+		// ordered:
 
 
 		int probeCount = rawData.length;
@@ -41,7 +42,7 @@ public class QuantileMathCommonsNormalization {
 		}
 
 		for (int probeID = 0; probeID < probeCount; probeID++) {
-			rankedMean[probeID] /= (double) sampleCount;
+			rankedMean[probeID] /= sampleCount;
 		}
 
 		double[] rankedMeanClasses = new double[probeCount - 1];
@@ -75,18 +76,18 @@ public class QuantileMathCommonsNormalization {
 	}
 
 	/**
-	 * Quantile normalize data where missing values are allowed. the missing values are replaced with the median or mean of the other probes in the sample If chosen keep median value instead of the NA value.
+	 * Quantile normalize data where missing values are allowed. the missing values are replaced with
+	 * the median or mean of the other probes in the sample If chosen keep median value instead of the
+	 * NA value.
 	 *
 	 * NB: For now NA is defined as "-999"!!
 	 *
-	 * @param rawData
-	 *            matrix containing expression/methylation data
-	 * @param useMean
-	 *            use mean for guessing the NA value if false use median
-	 * @param retainNA
-	 *            retain the NA values, put NA values back after normalization
+	 * @param rawData matrix containing expression/methylation data
+	 * @param useMean use mean for guessing the NA value if false use median
+	 * @param retainNA retain the NA values, put NA values back after normalization
 	 */
-	public static void QuantileNormAdressingNaValuesBeforeQN(double[][] rawData, boolean useMedian, boolean retainNA) {
+	public static void QuantileNormAdressingNaValuesBeforeQN(	double[][] rawData, boolean useMedian,
+																														boolean retainNA) {
 		boolean[][] wasNA = new boolean[rawData.length][rawData[1].length];
 
 		for (int s = 0; s < rawData[1].length; ++s) {
